Some updates for TUI
